Renames and moving CSS closer to components
