+++ src/app/move-popup/dimMoveItemProperties.html
