+++ src/app/move-popup/item-tag.scss
@@ -1,0 +1,7 @@
+dim-item-tag {
+  select {
+    background: transparent;
+    border: none;
+    outline: none;
+  }
+}
