+++ src/app/move-popup/move-popup.module.js
@@ -1,11 +1,11 @@
 import angular from 'angular';
-import { ItemStatsComponent } from './dimItemStats.component';
-import { ItemTagComponent } from './dimItemTag.directive';
-import { MoveAmount } from './dimMoveAmount.directive';
+import { ItemStatsComponent } from './item-stats.component';
+import { ItemTagComponent } from './item-tag.component';
+import { MoveAmount } from './move-amount.directive';
 import { ObjectivesComponent } from './objectives.component';
 import { MoveItemProperties } from './dimMoveItemProperties.directive';
 import { MovePopupComponent } from './dimMovePopup.directive';
-import { talentGridNodesFilter, TalentGrid } from './dimTalentGrid.directive';
+import { talentGridNodesFilter, TalentGridComponent } from './talent-grid.component';
 import { ItemPopup } from './item-popup.directive';
 import { PressTip } from './press-tip.directive';
 
@@ -17,7 +17,7 @@ export default angular
   .component('dimObjectives', ObjectivesComponent)
   .directive('dimMoveItemProperties', MoveItemProperties)
   .component('dimMovePopup', MovePopupComponent)
-  .directive('dimTalentGrid', TalentGrid)
+  .component('dimTalentGrid', TalentGridComponent)
   .filter('talentGridNodes', () => talentGridNodesFilter)
   .directive('itemPopup', ItemPopup)
   .directive('pressTip', PressTip)
